A_0026_001: Use shared_from_this to Player class

+++ Civ_Rob_2.0/src/UnitTemplate.cpp
@@ -83,7 +83,7 @@ UnitTemplate
 	LOG(R2D::LogLevelType::info, 0, logS::WHO::GAMEPLAY, logS::WHAT::LOAD_UNIT_CONFIG, logS::DATA::END);
 }
 
-unsigned int
+int64_t
 UnitTemplate
 ::searchUnitByName(const std::string& name)const
 {
